+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Set logger [#80](https://github.com/rokwire/notifications-building-block/issues/80)
+
 ## [1.4.0] - 2022-11-16
 ### Fixed
 - Fix docs path and the Dockerfile [#104](https://github.com/rokwire/notifications-building-block/issues/104)