+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Set logger [#80](https://github.com/rokwire/notifications-building-block/issues/80)
+- Rate limits [#98](https://github.com/rokwire/notifications-building-block/issues/98)
 
 ### Changed
 - Use auth library authorization [#75](https://github.com/rokwire/notifications-building-block/issues/75)