+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix API docs [#110](https://github.com/rokwire/notifications-building-block/issues/110)
+
 ### Added
 - Set logger [#80](https://github.com/rokwire/notifications-building-block/issues/80)